% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_delay_trends.R
\name{plot_delay_trends}
\alias{plot_delay_trends}
\title{Plot delay trends}
\usage{
plot_delay_trends(data, time_col, delay_col)
}
\arguments{
\item{data}{A data frame containing airline delay data.}

\item{time_col}{The name of the time column.}

\item{delay_col}{The name of the numeric delay column.}
}
\value{
A ggplot object showing delay trends.
}
\description{
Creates a line plot showing average delay over a selected time variable.
}
