% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_delays.R
\name{summarize_delays}
\alias{summarize_delays}
\title{Summarize airline delays}
\usage{
summarize_delays(data, airline_col, delay_col)
}
\arguments{
\item{data}{A data frame containing airline delay data.}

\item{airline_col}{The name of the airline column.}

\item{delay_col}{The name of the arrival delay column.}
}
\value{
A summary data frame with average delay by airline.
}
\description{
Calculates average arrival delay by airline.
}
